split interface
